Move lightning-fee use-case to wallets & rename
