+++ src/app/wallets/index.ts
@@ -11,6 +11,7 @@ export * from "./update-on-chain-receipt"
 export * from "./update-pending-invoices"
 export * from "./update-pending-payments"
 export * from "./add-invoice-for-wallet"
+export * from "./ln-probe-for-fee"
 export * from "./ln-send-payment"
 export * from "./intraledger-send-payment"
 export * from "../users/username-available"
